requirements.txt
ccxt>=3.0.0,<4.0.0
python-dotenv>=0.21.0,<2.0.0
